+++ R/weigh.R
@@ -3,7 +3,7 @@ NULL
 
 #' Apply Weight to Matrix
 #' 
-#' @param .Object the matrix to be weighed
+#' @param .Object A \code{matrix}, or a \code{count}-object.
 #' @param method the kind of weight to apply
 #' @param ... further parameters
 #' @exportMethod weigh
@@ -55,3 +55,61 @@ setMethod("weigh", "DocumentTermMatrix", function(.Object, method = "tfidf"){
   return(.Object)
 })
 
+
+
+#' @param with A \code{data.table} used to weigh p-attributes. A column 'weight' with term weights is
+#' required, and columns with the p-attributes of \code{.Object} for matching.
+#' @examples 
+#' \dontrun{
+#' library(data.table)
+#' if (require("zoo") && require("devtools")){
+#' 
+#' # Source in function 'get_sentiws' from a GitHub gist
+#' devtools::source_url("https://gist.githubusercontent.com/PolMine/70eeb095328070c18bd00ee087272adf/raw/c2eee2f48b11e6d893c19089b444f25b452d2adb/sentiws.R")
+#' SentiWS <- get_sentiws()
+#' 
+#' # Do the statistical word context analysis
+#' use("GermaParl")
+#' C <- context("GERMAPARL", query = "Islam", p_attribute = c("word", "pos"))
+#' PB <- partition_bundle(C, node = FALSE)
+#' name(PB) <- unlist(s_attributes(PB, s_attribute = "date"))
+#' PB <- weigh(PB, with = SentiWS)
+#' dt <- summary(PB)
+#' 
+#' # Aggregate by year
+#' dt[["year"]] <- gsub("^(\\d{4})-\\d{2}-\\d+$", "\\1-01-01", as.character(dt[["name"]]))
+#' dt_by_year <- aggregate(dt[,c("size", "positive_n", "negative_n")], list(dt[,"year"]), sum)
+#' colnames(dt_by_year)[1] <- "year"
+#' 
+#' # Shares are better than absolut counts 
+#' dt_by_year$negative_share <- dt_by_year$negative_n / dt_by_year$size
+#' dt_by_year$positive_share <- dt_by_year$positive_n / dt_by_year$size
+#' 
+#' # Turn it into zoo object, and plot it
+#' Z <- zoo(
+#'   x = dt_by_year[, c("positive_share", "negative_share")],
+#'   order.by = as.Date(dt_by_year[,"year"])
+#' )
+#' plot(
+#'   Z, ylab = "polarity", xlab = "year",
+#'   main = "Word context of 'Islam': Share of positive/negative vocabulary",
+#'   cex = 0.8,
+#'   cex.main = 0.8
+#' )
+#' 
+#' }
+#' }
+#' @rdname weigh-method
+setMethod("weigh", "count", function(.Object, with){
+  setkeyv(x = .Object@stat, cols = .Object@p_attribute)
+  setkeyv(x = with, cols = .Object@p_attribute)
+  with_min <- with[, c(.Object@p_attribute, "weight"), with = FALSE]
+  .Object@stat <- with_min[.Object@stat]
+  .Object
+})
+
+#' @rdname weigh-method
+setMethod("weigh", "count_bundle", function(.Object, with){
+  .Object@objects <- lapply(.Object@objects, function(x) weigh(x, with = with))
+  .Object
+})